Refactor: Modularized compiler
